Comments Dao done

+++ devops/src/main/java/com/cg/devops/dao/ReportDaoImpl.java
@@ -140,7 +140,7 @@ public class ReportDaoImpl implements ReportDao {
 			throw new ProgramException(ErrorMessagesUtils.MESSAGE4);
 		}
 		
-		return (int) jsonObject.get("total");
+		return (int) (long)jsonObject.get("total");
 	}
 
 }
